GrownLawfulFunction/requirements.txt
aiohttp==3.11.11
aiosqlite==0.20.0
alembic==1.14.0
asyncpg==0.30.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
celery==5.4.0
cryptography
fastapi==0.115.6
httpx==0.28.1
itsdangerous==2.2.0
joblib==1.4.2
lxml==5.3.0
numpy==2.2.1
pandas==2.2.3
passlib
pydantic==2.10.4
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
redis==5.2.1
scikit-learn==1.6.1
SQLAlchemy[asyncio]==2.0.36
stripe==11.4.0
tenacity==9.0.0
uvicorn[standard]==0.34.0
aiohttp
aiosqlite
alembic
asyncpg
bcrypt
beautifulsoup4
celery
cryptography
fastapi
httpx
itsdangerous
joblib
lxml
numpy
pandas
passlib[bcrypt]
pydantic
python-dotenv
python-jose[cryptography]
python-multipart
redis
scikit-learn
sqlalchemy[asyncio]
stripe
tenacity
uvicorn[standard]
